+++ docs/releasenotes.rst
@@ -13,7 +13,8 @@ Development Version
   Python; we produce separate real and imaginary arrays.
 - implement Julia (version >= 1) read code for ragged arrays
 - implement Maple read code for ragged arrays
-- improve Maple read code arrays.
+- implement IDL read code for ragged arrays
+- improve Maple read code arrays
 - implement flexible paths read code functions
 
 Version 0.4.1